Load tree with ext as root
This more closely matches how the VPK format is structured,
because there was good reason for it to be structured in that way!
Having Ext be at the root lets us get rid of various of the string
operations. It also makes it significantly cheaper to find what you
want, since you only have to look at a specific hashmap's keys.
This also swaps to indexmap, to make it easier to index
case-insensitiviely. Lots of data refers to materials/textures with
UPPER or lowercase (or unfortunately mixed case). A hashmap would force
you to either allocate a string, or iterate over the keys (doing string
comparisons! part of the point a hashmap is to hash!)