Do range based dir file instead
That this requires it to clone & store an `Arc<[u8]>` around on every
`DirFile` is annoying, though. It would be nicer if it could just be
passed in as a relevant parameter for deciding how to compare & hash
values.
I could do it via a custom fork of indexmap.
But I doubt the Arc is bottlenecking us too bad?